server/python/requirements.txt
celery~=5.2.7
kombu~=5.3.0b1
grpcio~=1.51.1
grpcio-tools~=1.51.1
prometheus-client~=0.15.0
protobuf~=4.21.9
py-grpc-prometheus~=0.7.0
